// Tracker/src/models/CryptoModel.ts
export interface TableData{
    id: number;
    icon: string;
    name: string;
    symbol: string;
    price:number;
    hour1: number;
    hour24: number;
    day7: number;
    marketCap: number;
    volumeInDollar: number;
    volumeInAsset: number;
    circulatingSupply: number;
    chart: string;
}